Move action entity extraction into a shared module
extract_entities_from_action_config walks action configs recursively: it
understands target blocks, service data and nested choose/if/parallel/repeat
steps, and keeps service names from being mistaken for entities. It lived in
the automation ectoplasm and was not reachable from anywhere else.

This moves it, together with extract_entities_from_value and its private
helpers, into custom_components/spook/action_extraction.py. The automation
repair imports the two public functions back and is otherwise untouched --
a pure relocation, no behaviour change.

reference_extraction was considered as the home for this, but it is
deliberately synchronous and hass-free and covers areas, devices, floors and
labels rather than entities. action_extraction sits next to
template_extraction instead, which has the same async, hass-aware,
entity-focused shape.
